requirements.txt
# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# Statistical analysis
scipy>=1.9.0
statsmodels>=0.13.0

# Copula analysis
copulas>=0.6.0
copulae>=0.1.0

# Machine learning
scikit-learn>=1.1.0

# Visualization
matplotlib>=3.5.0
plotly>=5.10.0

# API integration
pybit>=5.11.0

# Database
sqlite3  # Built-in with Python
psycopg2-binary>=2.9.0  # For PostgreSQL (future cloud migration)

# Configuration and utilities
python-dotenv>=0.19.0
pyyaml>=6.0

# Logging and monitoring
loguru>=0.6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0 